requirements.txt
crawl4ai[all]
python-dotenv
litellm
pydantic
langchain-text-splitters
pypdf
python-docx
sentence-transformers
chromadb
torch
torchvision
torchaudio
streamlit
plotly
networkx
streamlit-agraph
streamlit-ace
streamlit-option-menu
extra-streamlit-components